test for sehandler working. needs more tests.

+++ src/maidsafe/client/localstoremanager.cc
@@ -79,17 +79,25 @@ void ExecReturnLoadPacketCallback(const LoadPacketFunctor &cb,
 }
 
 LocalStoreManager::LocalStoreManager(
-    boost::shared_ptr<ChunkStore> client_chunkstore,
-    const boost::uint8_t k)
+    boost::shared_ptr<ChunkStore> client_chunkstore, const boost::uint8_t k)
         : K_(k),
           upper_threshold_(
               static_cast<boost::uint16_t>(K_ * kMinSuccessfulPecentageStore)),
-          db_(),
-          vbph_(),
-          mutex_(),
+          db_(), vbph_(), mutex_(),
           local_sm_dir_(file_system::LocalStoreManagerDir().string()),
           client_chunkstore_(client_chunkstore),
           ss_(SessionSingleton::getInstance()) {}
+
+LocalStoreManager::~LocalStoreManager() {
+#ifdef DEBUG
+  printf("Afore destruction\n");
+#endif
+  while (!chunks_pending_.empty())
+    boost::this_thread::sleep(boost::posix_time::seconds(1));
+#ifdef DEBUG
+  printf("After destruction\n");
+#endif
+}
 
 void LocalStoreManager::Init(int, kad::VoidFunctorOneString cb,
                              fs::path db_directory) {
@@ -147,13 +155,12 @@ int LocalStoreManager::LoadChunk(const std::string &chunk_name,
   return FindAndLoadChunk(chunk_name, data);
 }
 
-int LocalStoreManager::StoreChunk(const std::string &chunk_name,
-                                  const DirType,
+int LocalStoreManager::StoreChunk(const std::string &chunk_name, const DirType,
                                   const std::string&) {
-#ifdef LOCAL_PDVAULT
-  // Simulate knode lookup in AddToWatchList
+//  #ifdef LOCAL_PDVAULT
+//  // Simulate knode lookup in AddToWatchList
 //  boost::this_thread::sleep(boost::posix_time::seconds(2));
-#endif
+//  #endif
   std::string hex_chunk_name(base::EncodeToHex(chunk_name));
   fs::path file_path(local_sm_dir_ + "/StoreChunks");
   file_path = file_path / hex_chunk_name;
@@ -162,27 +169,45 @@ int LocalStoreManager::StoreChunk(const std::string &chunk_name,
   ChunkType type = client_chunkstore_->chunk_type(chunk_name);
   fs::path current = client_chunkstore_->GetChunkPath(chunk_name, type, false);
   try {
-    if (!fs::exists(file_path)) {
-      fs::copy_file(current, file_path);
+    if (fs::exists(current)) {
+      if (!fs::exists(file_path)) {
+        fs::copy_file(current, file_path);
+      }
+    } else {
+#ifdef DEBUG
+      printf("Chunk(%s) to store doesn't exist.\n", hex_chunk_name.c_str());
+#endif
+      signal_mutex_.lock();
+      chunks_pending_.insert(chunk_name);
+      signal_mutex_.unlock();
+      boost::thread thr(boost::bind(&LocalStoreManager::ExecuteReturnSignal,
+                                    this, chunk_name, kSendChunkFailure));
+      return kChunkStorePending;
     }
   }
   catch(const std::exception &e) {
 #ifdef DEBUG
     printf("%s\n", e.what());
+#endif
+    signal_mutex_.lock();
+    chunks_pending_.insert(chunk_name);
+    signal_mutex_.unlock();
     boost::thread thr(boost::bind(&LocalStoreManager::ExecuteReturnSignal, this,
                                   chunk_name, kSendChunkFailure));
     return kChunkStorePending;
-#endif
   }
+
   // Move chunk from Outgoing to Normal.
-  ChunkType chunk_type =
-      client_chunkstore_->chunk_type(chunk_name);
+  ChunkType chunk_type = client_chunkstore_->chunk_type(chunk_name);
   ChunkType new_type = chunk_type ^ (kOutgoing | kNormal);
   if (client_chunkstore_->ChangeChunkType(chunk_name, new_type) != 0) {
 #ifdef DEBUG
     printf("In LocalStoreManager::SendContent, failed to change chunk type.\n");
 #endif
   }
+  signal_mutex_.lock();
+  chunks_pending_.insert(chunk_name);
+  signal_mutex_.unlock();
   boost::thread thr(boost::bind(&LocalStoreManager::ExecuteReturnSignal, this,
                                 chunk_name, kSuccess));
   return kChunkStorePending;
@@ -1127,7 +1152,9 @@ void LocalStoreManager::ExecuteReturnSignal(const std::string &chunkname,
                                             ReturnCode rc) {
   int sleep_seconds((base::RandomInt32() % 5) + 1);
   boost::this_thread::sleep(boost::posix_time::seconds(sleep_seconds));
+  boost::mutex::scoped_lock loch_laggan(signal_mutex_);
   sig_chunk_uploaded_(chunkname, rc);
+  chunks_pending_.erase(chunkname);
 }
